Test `main` rather than `make_report`
The test is the same: We've just moved it from one module to another
module and passed it arguments by patching `sys.argv`. However, we
exercise more code.
